+++ CHANGELOG.md
@@ -26,6 +26,7 @@ with the exception that 0.x versions can break between minor versions.
 - Fix quadratic runtime when parsing pathological backticks
 - Fix StackOverflowError when parsing pathological HTML block attributes
 - Fix StackOverflowError when parsing pathological autolink email addresses
+- Fix StackOverflowError when parsing deeply nested emphasis
 
 ## [0.29.0] - 2026-06-20
 ### Added